Test: add test suite to ensure a RaftStorage impl works correctly
`openraft::testing::Suite::test_all(|| MemStore::new())` will run a full
test on every RaftStorage APIs with `MemStore`.

TODO: `testing::Suite` needs to test state machine APIs and snapshot APIs.
